# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TeamPulse API",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/tests": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Список тестов",
                "operationId": "2d6527e0e425c66092b7ce936622f79e",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество записей",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Смещение",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total": {
                                            "description": "Общее количество записей",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Test"
                ],
                "summary": "Создать тест",
                "operationId": "43c9b4110acb891d5bdcb0eef84a0d88",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/tests/templates": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Список шаблонов",
                "operationId": "2a9ae385076e308863aa810ac779f5d3",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество записей",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Смещение",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total": {
                                            "description": "Общее количество записей",
                                            "type": "integer"
                                        },
                                        "tests": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TestTemplateResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tests/{uuid}/users": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Список назначенных пользователей",
                "operationId": "d7342da7efa1ea7b65ab34f09a7f9ecd",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество записей",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Смещение",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total": {
                                            "description": "Общее количество записей",
                                            "type": "integer"
                                        },
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tests/{uuid}/groups": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Список назначенных групп",
                "operationId": "2f7985edb2c8a0c49ddf83b6f146a658",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество записей",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Смещение",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total": {
                                            "description": "Общее количество записей",
                                            "type": "integer"
                                        },
                                        "groups": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Group"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/tests/{uuid}/topics/{topicUuid}": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Получить вопросы теста по ID топика",
                "operationId": "7bf0fe72ce73cf5542140c977bf3397c",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество записей",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Смещение",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total": {
                                            "description": "Общее количество записей",
                                            "type": "integer"
                                        },
                                        "questions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/QuestionsResponse"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tests/{uuid}/assign": {
            "post": {
                "tags": [
                    "Test"
                ],
                "summary": "Назначить тест",
                "operationId": "ef5f3186ad1e4c1b36e2435b446d0920",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignTestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AnswerType": {
                "description": "Типы ответов\n *      0 - Текст\n *      1 - Одиночный выбор\n *      2 - Множественный выбор\n *      3 - Оценка\n *      4 - Распределение баллов",
                "type": "integer",
                "enum": [
                    "Text",
                    "SingleChoice",
                    "MultipleChoice",
                    "Scale"
                ],
                "example": 0
            },
...
            "TestStatus": {
                "description": "Статус теста\n *          draft - Черновик\n *          done - Готовый тест",
                "type": "string",
                "enum": [
                    "Draft",
                    "Done"
                ],
                "example": "draft"
            },
...
            "AssignTestRequest": {
                "description": "Тело запроса для назначения теста",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "New test name"
                    },
                    "description": {
                        "type": "string",
                        "example": "New test description"
                    },
                    "frequency": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "startDate": {
                        "description": "Временная метка начала прохождения теста",
                        "type": "string",
                        "format": "datetime"
                    },
                    "endDate": {
                        "description": "Временная метка окончания прохождения теста",
                        "type": "string",
                        "format": "datetime"
                    },
                    "subjectId": {
                        "description": "ID пользователя, на оценку которого направлен тест",
                        "type": "string",
                        "format": "uuid"
                    },
                    "assignToAll": {
                        "description": "Метка назначения всем сотрудникам",
                        "type": "boolean"
                    },
                    "groups": {
                        "type": "array",
                        "items": {
                            "description": "ID группы",
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "employees": {
                        "type": "array",
                        "items": {
                            "description": "ID пользователя",
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    },
                    "isAnonymous": {
                        "description": "Метка анонимности",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "CreateTestRequest": {
                "description": "Тело запроса для создания теста",
                "properties": {
                    "name": {
                        "description": "Название теста",
                        "type": "string",
                        "example": "New test"
                    },
                    "description": {
                        "description": "Описание теста",
                        "type": "string",
                        "example": "New test description"
                    },
                    "status": {
                        "$ref": "#/components/schemas/TestStatus"
                    },
                    "tests": {
                        "description": "Список вопросов теста",
                        "type": "array",
                        "items": {
                            "properties": {
                                "topic": {
                                    "description": "Тема вопроса",
                                    "type": "string",
                                    "example": "default"
                                },
                                "questions": {
                                    "description": "Список вопросов",
                                    "type": "array",
                                    "items": {
                                        "required": [
                                            "name",
                                            "type"
                                        ],
                                        "properties": {
                                            "name": {
                                                "description": "Текст вопроса",
                                                "type": "string",
                                                "example": "Question 1"
                                            },
                                            "type": {
                                                "description": "Тип ответа на вопрос",
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "tags": {
                                                "description": "Список тегов",
                                                "type": "array",
                                                "items": {
                                                    "required": [
                                                        "name",
                                                        "points"
                                                    ],
                                                    "properties": {
                                                        "name": {
                                                            "description": "Тег вопроса",
                                                            "type": "string",
                                                            "example": "Tag 1"
                                                        },
                                                        "points": {
                                                            "description": "Количество поинтов на тег",
                                                            "type": "integer",
                                                            "example": 8
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "answers": {
                                                "description": "Список ответов",
                                                "type": "array",
                                                "items": {
                                                    "required": [
                                                        "text"
                                                    ],
                                                    "properties": {
                                                        "text": {
                                                            "description": "Текст ответа",
                                                            "type": "string",
                                                            "example": "Answer 1"
                                                        },
                                                        "points": {
                                                            "description": "Список поинтов для ответа",
                                                            "type": "array",
                                                            "items": {
                                                                "required": [
                                                                    "name",
                                                                    "points"
                                                                ],
                                                                "properties": {
                                                                    "name": {
                                                                        "description": "Название тега",
                                                                        "type": "string",
                                                                        "example": "Point 1"
                                                                    },
                                                                    "points": {
                                                                        "description": "Количество баллов",
                                                                        "type": "integer",
                                                                        "example": 5
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "TestTemplateResource": {
                "description": "Шаблон теста",
                "properties": {
                    "id": {
                        "description": "ID теста",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Название теста",
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/TestStatus"
                    },
                    "author": {
                        "$ref": "#/components/schemas/User"
                    },
                    "createdAt": {
                        "description": "Дата создания",
                        "type": "string"
                    },
                    "updatedAt": {
                        "description": "Дата обновления",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "AnswerTagShort": {
                "properties": {
                    "tagId": {
                        "description": "ID тега",
                        "type": "string",
                        "format": "uuid"
                    },
                    "points": {
                        "description": "Количество баллов",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
            "Group": {
                "description": "Группа",
                "properties": {
                    "id": {
                        "description": "ID группы",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Название группы",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "QuestionsResponse": {
                "description": "Вопрос",
                "properties": {
                    "id": {
                        "description": "ID вопроса",
                        "type": "string",
                        "format": "uuid"
                    },
                    "text": {
                        "description": "Текст вопроса",
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/AnswerType"
                    },
                    "tags": {
                        "description": "Теги вопроса",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    },
                    "answers": {
                        "description": "Теги вопроса",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "ID ответа",
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "text": {
                                    "description": "Текст ответа",
                                    "type": "string"
                                },
                                "tagPoints": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnswerTagShort"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "Tag": {
                "description": "Тег",
                "properties": {
                    "id": {
                        "description": "ID тега",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Название тега",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "description": "Пользователь",
                "properties": {
                    "id": {
                        "description": "ID пользователя",
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "description": "Имя",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "Фамилия",
                        "type": "string"
                    },
                    "login": {
                        "description": "Логин",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string"
                    },
                    "role": {
                        "description": "Роль",
                        "type": "string"
                    },
                    "status": {
                        "description": "Статус сущности",
                        "type": "string"
                    },
                    "createdAt": {
                        "description": "Дата создания",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}